refactoring #9
