+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/HttpClientFactory.java
@@ -28,6 +28,7 @@ import java.util.Optional;
 import java.util.Random;
 import java.util.concurrent.ExecutorService;
 
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 
 /**
