+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/okhttp/authentication/AuthenticationsConfigurer.java
@@ -13,7 +13,8 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
-import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.PROXY_PREFIX;
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.PROXY_PREFIX;
+
 
 /**
  * configure authentication settings for {@link OkHttpClient}.
