+++ kafka-connect-http-connectors/src/main/java/io/github/clescot/kafka/connect/http/client/proxy/ProxySelectorFactory.java
@@ -9,6 +9,7 @@ import java.util.*;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
 
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 
 public class ProxySelectorFactory {
