+++ kafka-connect-http-connectors/src/test/java/io/github/clescot/kafka/connect/http/client/okhttp/authentication/DigestAuthenticationConfigurerTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import java.util.Map;
 import java.util.Random;
 
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 import static org.assertj.core.api.Assertions.assertThat;
