+++ kafka-connect-http-connectors/src/test/java/io/github/clescot/kafka/connect/http/client/okhttp/authentication/OAuth2ClientCredentialsFlowAuthenticatorTest.java
@@ -22,6 +22,7 @@ import java.util.Map;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.stubbing.Scenario.STARTED;
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
