+++ kafka-connect-http-connectors/src/test/java/io/github/clescot/kafka/connect/http/sink/HttpSinkTaskTest.java
@@ -64,6 +64,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.stubbing.Scenario.STARTED;
 import static io.github.clescot.kafka.connect.http.client.config.HttpRequestPredicateBuilder.URL_REGEX;
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.mapper.HttpRequestMapperFactory.JEXL_ALWAYS_MATCHES;
 import static org.assertj.core.api.Assertions.assertThat;
