+++ kafka-connect-http-connectors/src/test/java/io/github/clescot/kafka/connect/http/sink/publish/PublishConfigurerTest.java
@@ -19,6 +19,7 @@ import org.junit.jupiter.api.Test;
 import java.util.HashMap;
 
 import static io.github.clescot.kafka.connect.http.core.queue.QueueFactory.DEFAULT_QUEUE_NAME;
+import static io.github.clescot.kafka.connect.http.sink.HttpClientConfigDefinition.*;
 import static io.github.clescot.kafka.connect.http.sink.HttpConfigDefinition.*;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
